Refactorizes github config classes
