+++ core/src/main/scala/sbtorgpolicies/github/GitHubOps.scala
@@ -22,17 +22,17 @@ import cats.data.{EitherT, NonEmptyList}
 import cats.free.Free
 import cats.implicits._
 import cats.syntax.either._
-import com.github.marklister.base64.Base64.Encoder
+import com.github.marklister.base64.Base64._
 import github4s.Github
 import github4s.GithubResponses._
 import github4s.free.domain._
 import sbt.IO
 import sbtorgpolicies.exceptions.{GitHubException, IOException, OrgPolicyException}
+import sbtorgpolicies.github.config._
 import sbtorgpolicies.github.instances._
 import sbtorgpolicies.github.syntax._
 import sbtorgpolicies.io.syntax._
 import sbtorgpolicies.io.{FileReader, IOResult}
-import com.github.marklister.base64.Base64._
 
 class GitHubOps(owner: String, repo: String, accessToken: Option[String]) {
